--- tests/dummy/app/routes/actionable.ts
@@ -1,3 +1,0 @@
-import Route from '@ember/routing/route';
-
-export default class Actionable extends Route {}
