+++ tests/dummy/app/routes/input.ts
@@ -1,6 +1,6 @@
 import Route from '@ember/routing/route';
 
-export default class Scrollable extends Route.extend({
+export default class Input extends Route.extend({
   // anything which *must* be merged to prototype here
 }) {
   // normal class body definition here
