+++ tests/dummy/app/routes/visual.ts
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class Visual extends Route {}
